Crixus first lines in Cuda, no makefile yet.

+++ src/crixus.h
@@ -1,9 +1,9 @@
 #define DATASETNAME "Compound"
-#define sgn(x) (double)((x>0.)-(x<0.))
+#define sgn(x) (float)((x>0.)-(x<0.))
 
 struct OutBuf{
-	double x,y,z,nx,ny,nz,vol,surf;
+	float x,y,z,nx,ny,nz,vol,surf;
 	int kpar, kfluid, kent, kparmob, iref, ep1, ep2, ep3;
 };
 
-int hdf5_output (OutBuf *buf, int len, const char *filename, double *timevalue);
+int hdf5_output (OutBuf *buf, int len, const char *filename, float *timevalue);
